Tag v1.6.0

+++ 5G_Core_Network/full_uninstall.sh
@@ -60,15 +60,21 @@ echo "Removing Open5GS installation directory..."
 sudo rm -rf open5gs/
 sudo rm -rf /var/log/open5gs
 
-for INTERMEDIATE_DIR in open5gs-*; do
-    if [[ -d "$INTERMEDIATE_DIR" && "$INTERMEDIATE_DIR" != "open5gs-*" ]]; then
-        echo "Removing intermediate open5gs directory: $INTERMEDIATE_DIR"
-        sudo rm -rf "$INTERMEDIATE_DIR"
-    fi
-done
+if ls open5gs-* 1>/dev/null 2>&1; then
+    echo "Removing intermediate open5gs directories from older versions..."
+    for INTERMEDIATE_DIR in open5gs-*; do
+        if [[ -d "$INTERMEDIATE_DIR" && "$INTERMEDIATE_DIR" != "open5gs-*" ]]; then
+            echo "Removing intermediate open5gs directory: $INTERMEDIATE_DIR"
+            sudo rm -rf "$INTERMEDIATE_DIR"
+        fi
+    done
+fi
 
+mkdir -p logs
+cd logs
 echo "Uninstalling WebUI..."
 curl -fsSL https://open5gs.org/open5gs/assets/webui/uninstall | sudo -E bash -
+cd ..
 
 echo "Performing general system cleanup..."
 sudo apt-get autoremove -y
